kafka_consumer/requirements.txt
kafka-python-ng==2.2.3
psycopg2-binary==2.9.10
requests==2.32.3

